requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.3.0
torchvision==0.18.0
#segment_anything
transformers>=4.25.1
Pillow
opencv-python
rembg
argparse
git+https://github.com/facebookresearch/segment-anything.git
